messing around with getqueue placement
