+++ src/utils/getLocalCommands.js
@@ -15,7 +15,7 @@ module.exports = (exceptions = []) =>{
 
         for(const commandFile of commandFiles){
             const commandObject = require(commandFile);
-
+        
             if(exceptions.includes(commandObject.data.name))
                 continue;
